[Refactor RDK-B WiFi HAL]
[Description]
Refactor wifi_getRadioPossibleChannels. Use iwlist to parse possible channel.
Set buffer size as 128 bytes rather than 64 bytes (limit of document).
The reason is that the result of 5GHz possible channel is over than 64 bytes. If we set 64 bytes the result would be cut off by the size limitation.
And the caller, WiFiAgent and TR181, also give 128 bytes to the buffer, so set 128 bytes would cause error.

[Release-log]
N/A